# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Libre-Q API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/recruitment/{slug}/links": {
      "get": {
        "tags": [
          "admin-recruitment",
          "Admin Recruitment"
        ],
        "summary": "List Study Links",
        "description": "List all recruitment links for a specific study.",
        "operationId": "list_study_links_api_admin_recruitment__slug__links_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecruitmentLinkRead"
                  },
                  "title": "Response List Study Links Api Admin Recruitment  Slug  Links Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-recruitment",
          "Admin Recruitment"
        ],
        "summary": "Create Recruitment Links",
        "description": "Create one or more recruitment links.",
        "operationId": "create_recruitment_links_api_admin_recruitment__slug__links_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Count"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecruitmentLinkCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecruitmentLinkRead"
                  },
                  "title": "Response Create Recruitment Links Api Admin Recruitment  Slug  Links Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/recruitment/links/{link_id}": {
      "delete": {
        "tags": [
          "admin-recruitment",
          "Admin Recruitment"
        ],
        "summary": "Revoke Recruitment Link",
        "description": "Revoke a recruitment link.",
        "operationId": "revoke_recruitment_link_api_admin_recruitment_links__link_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Link Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecruitmentLinkCreate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/RecruitmentLinkType",
            "default": "public"
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          }
        },
        "type": "object",
        "title": "RecruitmentLinkCreate",
        "description": "Schema for creating a recruitment link."
      },
      "RecruitmentLinkRead": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/RecruitmentLinkType",
            "default": "public"
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "study_id": {
            "type": "integer",
            "title": "Study Id"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "usage_count": {
            "type": "integer",
            "title": "Usage Count"
          },
          "start_count": {
            "type": "integer",
            "title": "Start Count"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "study_id",
          "token",
          "usage_count",
          "start_count",
          "created_at"
        ],
        "title": "RecruitmentLinkRead",
        "description": "Schema for reading a recruitment link."
      },
      "RecruitmentLinkType": {
        "type": "string",
        "enum": [
          "public",
          "individual",
          "limited"
        ],
        "title": "RecruitmentLinkType",
        "description": "Enum for types of recruitment links."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/token"
          }
        }
      }
    }
  }
}